fix(security): vibe code audit — 5 vulnerabilities + 50+ code quality fixes
Security fixes:
- fix(api-gateway): WebSocket origin check bypassed when CORS_ORIGINS unset
- fix(web): XSS via identity-function sanitizer stubs in utils/security.ts
- fix(web): XSS via unsanitized Mermaid SVG in MermaidRenderer
- fix(web): XSS via unsanitized KaTeX HTML in CalculationCard
- fix(desktop): SQL format string in maintenance.rs → static literals
- fix(desktop): BrowserDebugTabs regex sanitization → DOMPurify

Code quality:
- fix(web): 8 empty catch blocks → proper error logging
- fix(mobile): 10 empty catch blocks → proper error logging
- fix(extension): 13 empty catch blocks → proper error logging
- fix(extension-vscode): 10 empty catch blocks → proper error logging
- fix(desktop-rust): 2 production unwrap() → proper error propagation
- fix(desktop-rust): 5 clippy warnings resolved (cast, iter, spawn)
- fix(desktop): 8 competitor reference comments cleaned up
- fix(web): sanitizeSVG extended with missing SVG attributes

+++ apps/desktop/src-tauri/src/core/agent/autonomous.rs
@@ -329,7 +329,7 @@ impl AutonomousAgent {
             // Enforce cumulative cost cap (same as run_autonomous_loop)
             if let Ok(router) = self.router.try_read() {
                 let cost = router.get_cumulative_cost();
-                if cost > self.config.max_session_cost as f64 {
+                if cost > self.config.max_session_cost {
                     return Err(anyhow!(
                         "run_goal exceeded session cost cap (${:.2} > ${:.2})",
                         cost,
